requirements.txt
attrs==21.2.0
aws-cdk.assets==1.135.0
aws-cdk.aws-acmpca==1.135.0
aws-cdk.aws-apigateway==1.135.0
aws-cdk.aws-applicationautoscaling==1.135.0
aws-cdk.aws-autoscaling==1.135.0
aws-cdk.aws-autoscaling-common==1.135.0
aws-cdk.aws-autoscaling-hooktargets==1.135.0
aws-cdk.aws-certificatemanager==1.135.0
aws-cdk.aws-cloudformation==1.135.0
aws-cdk.aws-cloudfront==1.135.0
aws-cdk.aws-cloudwatch==1.135.0
aws-cdk.aws-codebuild==1.135.0
aws-cdk.aws-codecommit==1.135.0
aws-cdk.aws-codedeploy==1.135.0
aws-cdk.aws-codeguruprofiler==1.135.0
aws-cdk.aws-codepipeline==1.135.0
aws-cdk.aws-codepipeline-actions==1.135.0
aws-cdk.aws-codestarnotifications==1.135.0
aws-cdk.aws-cognito==1.135.0
aws-cdk.aws-ec2==1.135.0
aws-cdk.aws-ecr==1.135.0
aws-cdk.aws-ecr-assets==1.135.0
aws-cdk.aws-ecs==1.135.0
aws-cdk.aws-efs==1.135.0
aws-cdk.aws-elasticloadbalancing==1.135.0
aws-cdk.aws-elasticloadbalancingv2==1.135.0
aws-cdk.aws-events==1.135.0
aws-cdk.aws-events-targets==1.135.0
aws-cdk.aws-globalaccelerator==1.135.0
aws-cdk.aws-iam==1.135.0
aws-cdk.aws-kinesis==1.135.0
aws-cdk.aws-kinesisfirehose==1.135.0
aws-cdk.aws-kms==1.135.0
aws-cdk.aws-lambda==1.135.0
aws-cdk.aws-logs==1.135.0
aws-cdk.aws-route53==1.135.0
aws-cdk.aws-route53-targets==1.135.0
aws-cdk.aws-s3==1.135.0
aws-cdk.aws-s3-assets==1.135.0
aws-cdk.aws-sam==1.135.0
aws-cdk.aws-secretsmanager==1.135.0
aws-cdk.aws-servicediscovery==1.135.0
aws-cdk.aws-signer==1.135.0
aws-cdk.aws-sns==1.135.0
aws-cdk.aws-sns-subscriptions==1.135.0
aws-cdk.aws-sqs==1.135.0
aws-cdk.aws-ssm==1.135.0
aws-cdk.aws-stepfunctions==1.135.0
aws-cdk.cloud-assembly-schema==1.135.0
aws-cdk.core==1.135.0
aws-cdk.custom-resources==1.135.0
aws-cdk.cx-api==1.135.0
aws-cdk.pipelines==1.135.0
aws-cdk.region-info==1.135.0
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.9
constructs==3.3.165
idna==3.3
iniconfig==1.1.1
jsii==1.48.0
packaging==21.3
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
